refactor: 자동차 경주 결과의 가장 높은 플레이어를 구하는 책임을 CarResult에게 할당
